Include node ports in hash
When using a non-standard port, I found that $port was not being set correctly and meshchatsync was using port 80 always. The pi_node_list subroutine only sets the 'platform' and 'node' keys in the hash references based on the port number, but it doesn't include the 'port' key in the hash reference.